// web/src/generated/SearchResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FacetData } from "./FacetData";
import type { SearchAccount } from "./SearchAccount";
import type { SearchRow } from "./SearchRow";

export type SearchResponse = { results: Array<SearchRow>, total: number, cross_account: boolean, facets: FacetData, accounts: Array<SearchAccount>, };