// web/src/generated/SearchRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Addr } from "./Addr";

export type SearchRow = { id: number, account_id: number, account: string, subject: string, subject_raw: string, from: Array<Addr>, date: number, date_offset_mins: number, date_source: string, skew: boolean, has_attach: boolean, thread_id: string, thread_size: number, folders: Array<string>, snippet: string, score: number, };