// web/src/generated/FacetData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExtFacet } from "./ExtFacet";
import type { OrgFacet } from "./OrgFacet";
import type { SenderFacet } from "./SenderFacet";
import type { YearFacet } from "./YearFacet";

export type FacetData = { senders: Array<SenderFacet>, orgs: Array<OrgFacet>, years: Array<YearFacet>, exts: Array<ExtFacet>, };